requirements.txt
fastapi
pydantic
pydantic[email]
uvicorn
httpx
# cache libraries
fastapi-cache2[redis]
aioredis